requirements.app.txt
Flask-GraphQL==1.1.0
GoogleAppEngineCloudStorageClient==1.9.22.1
graphene==0.7.3
